refactor: promote ConditionContext as public API and remove ConditionHelper content to RuleConfig
