--- liquibase-linter-core/src/test/java/io/github/liquibaselinter/ConditionHelperTest.java
@@ -1,88 +1,0 @@
-package io.github.liquibaselinter;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-
-import io.github.liquibaselinter.config.RuleConfig;
-import liquibase.ContextExpression;
-import liquibase.changelog.ChangeSet;
-import liquibase.changelog.DatabaseChangeLog;
-import org.junit.jupiter.api.Nested;
-import org.junit.jupiter.api.Test;
-
-class ConditionHelperTest {
-
-    @Nested
-    class SecurityHardening {
-
-        // Object.getClass() is outside the RESTRICTED+liquibase/io.github.liquibaselinter allowlist, so JEXL
-        // denies it (returning null rather than throwing); the ensuing failure to unbox to boolean is what
-        // ultimately stops the expression from ever resolving to a usable Class/Runtime reference.
-        @Test
-        void shouldDenyReflectionOutsideAllowedPackages() {
-            ChangeSet changeSet = changeSetWithNoContext();
-            RuleConfig ruleConfig = RuleConfig.builder()
-                .withCondition("changeSet.getClass().forName('java.lang.Runtime')")
-                .build();
-
-            assertThatThrownBy(() -> ConditionHelper.evaluateCondition(ruleConfig, changeSet)).isInstanceOf(
-                RuntimeException.class
-            );
-        }
-    }
-
-    @Nested
-    class MatchesContext {
-
-        @Test
-        void shouldMatchSimpleContext() {
-            ChangeSet changeSet = changeSetWithContext("foo");
-            RuleConfig ruleConfig = RuleConfig.builder().withCondition("matchesContext('foo')").build();
-
-            assertThat(ConditionHelper.evaluateCondition(ruleConfig, changeSet)).isTrue();
-        }
-
-        @Test
-        void shouldNotMatchNegativeSimpleContext() {
-            ChangeSet changeSet = changeSetWithContext("!foo");
-            RuleConfig ruleConfig = RuleConfig.builder().withCondition("matchesContext('foo')").build();
-
-            assertThat(ConditionHelper.evaluateCondition(ruleConfig, changeSet)).isFalse();
-        }
-
-        @Test
-        void shouldNotNatchSimpleContextMismatch() {
-            ChangeSet changeSet = changeSetWithContext("foo");
-            RuleConfig ruleConfig = RuleConfig.builder().withCondition("matchesContext('bar')").build();
-
-            assertThat(ConditionHelper.evaluateCondition(ruleConfig, changeSet)).isFalse();
-        }
-
-        @Test
-        void shouldNotMatchNoContext() {
-            ChangeSet changeSet = changeSetWithNoContext();
-            RuleConfig ruleConfig = RuleConfig.builder().withCondition("matchesContext('bar')").build();
-
-            assertThat(ConditionHelper.evaluateCondition(ruleConfig, changeSet)).isFalse();
-        }
-
-        @Test
-        void shouldMatchMultipleAndContexts() {
-            ChangeSet changeSet = changeSetWithContext("foo and bar");
-            RuleConfig ruleConfig = RuleConfig.builder().withCondition("matchesContext('foo', 'bar')").build();
-
-            assertThat(ConditionHelper.evaluateCondition(ruleConfig, changeSet)).isTrue();
-        }
-    }
-
-    private static ChangeSet changeSetWithContext(String contextExpression) {
-        ChangeSet changeSet = changeSetWithNoContext();
-        changeSet.setContextFilter(new ContextExpression(contextExpression));
-
-        return changeSet;
-    }
-
-    private static ChangeSet changeSetWithNoContext() {
-        return new ChangeSet(new DatabaseChangeLog());
-    }
-}
